move duplicates
