+++ skbase/utils/dependencies/tests/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Testing utilities for depeendency checkers."""
